Added an additional mode to SpeedInterfaceCounterAnims (from Mr.Stalin)
Refactored the code of Stats.cpp.
Unified stat_level_ usages in Inventory.cpp, LoadGameHook.cpp and skills.cpp.

+++ sfall/Stats.h
@@ -24,5 +24,5 @@ void _stdcall SetPCStatMax(int stat, int i);
 void _stdcall SetPCStatMin(int stat, int i);
 void _stdcall SetNPCStatMax(int stat, int i);
 void _stdcall SetNPCStatMin(int stat, int i);
-extern unsigned short StandardApAcBonus;
-extern unsigned short ExtraApAcBonus;
+extern DWORD StandardApAcBonus;
+extern DWORD ExtraApAcBonus;
